requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
blis==0.7.11
catalogue==2.0.10
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.19.0
colorama==0.4.6
confection==0.1.5
cymem==2.0.8
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
filelock==3.16.1
fsspec==2024.9.0
grpcio==1.66.1
idna==3.10
Jinja2==3.1.4
langcodes==3.4.1
language_data==1.2.0
marisa-trie==1.2.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.10
networkx==3.3
numpy==1.26.4
packaging==24.1
pandas==2.2.3
pillow==10.4.0
preshed==3.0.9
protobuf==5.28.2
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
rich==13.8.1
setuptools==75.1.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
spacy==3.7.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
sympy==1.13.3
tensorboard==2.17.1
tensorboard-data-server==0.7.2
thinc==8.2.5
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
utils==1.0.2
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.0.4
wrapt==1.16.0